fix: implement keyup handling in App class component
