--- src/components/Keyboard.tsx
@@ -1,39 +1,0 @@
-import { Component } from 'react';
-
-type State = {
-  pressedKey: string | null;
-};
-
-class Keyboard extends Component<{}, State> {
-  state: State = {
-    pressedKey: null,
-  };
-
-  componentDidMount() {
-    window.addEventListener('keydown', this.handleKeyDown);
-  }
-
-  componentWillUnmount() {
-    window.removeEventListener('keydown', this.handleKeyDown);
-  }
-
-  handleKeyDown = (event: KeyboardEvent) => {
-    this.setState({ pressedKey: event.key });
-  };
-
-  render() {
-    const { pressedKey } = this.state;
-
-    return (
-      <div className="Keyboard">
-        <p className="App__message">
-          {pressedKey === null
-            ? 'Nothing was pressed yet'
-            : `The last pressed key is [${pressedKey}]`}
-        </p>
-      </div>
-    );
-  }
-}
-
-export { Keyboard };
